+++ CHANGES.md
@@ -3,6 +3,17 @@ twilio-python Changelog
 
 Here you can see the full list of changes between each twilio-python release.
 
+[2026-05-07] Version 9.10.9
+---------------------------
+**Conversations**
+- update actions api visibility
+
+**Memory**
+- ## 2026-05-07
+- **Content updates**:
+- include store api
+
+
 [2026-05-07] Version 9.10.7
 ---------------------------
 **Library - Chore**